app/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
drf-yasg==1.21.5
psycopg2==2.9.5
uvicorn==0.20.0
